Update FacilityController
- Prevent a facility from being deleted if it has an open/pending update
request.

+++ api/app/Http/Controllers/FacilityController.php
@@ -97,6 +97,13 @@ class FacilityController extends Controller
     public function destroy(FacilityRequest $request, $id)
     {
         $f = Facility::findOrFail($id);
+
+        // Not allowed to delete a facility if it has an open/pending update
+        // request.
+        if ($f->currentRevision()->first()->fulB()->notClosed()->first()) {
+            abort(400);
+        }
+        
         $deletedFacility = $this->toCcArray($f->toArray());
         $f->delete();
         return $deletedFacility;
